Move views to model repo
