+++ src/main/java/com/rackspace/salus/monitor_management/web/controller/MonitorApiController.java
@@ -27,7 +27,7 @@ import com.rackspace.salus.monitor_management.web.model.BoundMonitorDTO;
 import com.rackspace.salus.monitor_management.web.model.DetailedMonitorInput;
 import com.rackspace.salus.monitor_management.web.model.DetailedMonitorOutput;
 import com.rackspace.salus.monitor_management.web.model.ValidationGroups;
-import com.rackspace.salus.monitor_management.web.model.View;
+import com.rackspace.salus.telemetry.model.View;
 import com.rackspace.salus.telemetry.model.NotFoundException;
 import com.rackspace.salus.telemetry.model.PagedContent;
 import io.swagger.annotations.Api;
