+++ src/main/java/com/rackspace/salus/monitor_management/web/model/ZoneDTO.java
@@ -16,6 +16,7 @@
 package com.rackspace.salus.monitor_management.web.model;
 
 import com.fasterxml.jackson.annotation.JsonView;
+import com.rackspace.salus.telemetry.model.View;
 import lombok.Data;
 
 import java.util.List;
