+++ src/test/java/com/rackspace/salus/monitor_management/web/model/ZoneDTOTest.java
@@ -23,6 +23,7 @@ import static org.junit.Assert.assertThat;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.rackspace.salus.monitor_management.entities.Zone;
+import com.rackspace.salus.telemetry.model.View;
 import org.junit.Test;
 import uk.co.jemos.podam.api.PodamFactory;
 import uk.co.jemos.podam.api.PodamFactoryImpl;
